Cleaned up files and comments for easier GitHub use

+++ clean.py
@@ -5,13 +5,13 @@ import numpy as np
 import os
 from collections import defaultdict as dd
 
-    
+# Return the correct area of the pitch given a location (meters x meters).
+# NOTE: I chose not to divide the pitch into equal boxes, but instead take into account the different areas of the pitch.
+# The biggest reason for this was the penalty boxes. I wanted to make those their own boxes, as the mentality signifcantly
+# changes when a player crosses into the penalty area.
 def square(location):
     x = location[0]
     y = location[1]
-    #I chose not to divide the pitch into equal boxes, but instead take into account the different areas of the pitch.
-    #The biggest reason for this was the 18 yard boxes. I wanted to make those their own boxes, as the mentality changes when a player
-    #crosses into the penalty area.
     if x <= 18:
         if y <= 18:
             return 1
@@ -71,21 +71,8 @@ def alt(key):
     else:
         return 0
 
-def pressure(press):
-    #Storing if a pass was made under pressure
-    #Note - I decided to not include this in the pass types as if a pass is under pressure, every option will be under pressure, which
-    #means that this factor does not have a signficant affect on which pass is a better choice, though I decided to keep the data
-    # for any possible future use
-    if type(press) == float:
-        if press == 1.0:
-            return 1
-        else:
-            return 0
-    else:
-        return 2
-
+# Finding if the pass was received under pressure and storing that for later use
 def receipt(x):
-    #finding if the pass was received under pressure and storing that for later use
     if x['type'] == {'id': 42, 'name': 'Ball Receipt*'}:
         if 'under_pressure' in x:
             if pd.isna(x['under_pressure']):
@@ -96,8 +83,8 @@ def receipt(x):
             store[x['id']] = [x['index'], 0]
     pass
 
+# Using the stored pressure data from the above function, finds the related ball receipt for each event and stores if it was under pressure
 def related_pressure(x):
-    #Using the stored pressure data from the above function, finds the related ball receipt for each event and stores if it was under pressure
     idx = x['index']
     if x['type'] == {'id': 30, 'name': 'Pass'}:
         if type(x['related_events']) == list:
@@ -114,8 +101,8 @@ def related_pressure(x):
     else:
         return 2
 
+# Saving the xg for each possession's first shot if there was a shot for later use
 def shots(event):
-    #Saving the xg for each possession's first shot if there was a shot for later use
     if event['type'] == {'id': 16, 'name': 'Shot'}:
         key = (event['possession'], event['possession_team']['id'], event['index'])
         if key not in shot.keys():
@@ -123,21 +110,21 @@ def shots(event):
             shot[key] = event['shot']['statsbomb_xg']
     pass
 
+# Storing if a pass was successful
 def success(p):
-    #Storing if a pass was successful
     if type(p) == dict:
         if 'outcome' not in p:
             return 1
         else:
             return 0
 
+# Finding the xg of each pass, and tracking only the events that we need for later use
 def xg(event):
-    #Finding the xg of each pass, and tracking only the events that we need for later use
     if type(event['pass']) == dict:
         if 'outcome' in event['pass']:
-            #Does not inlcude injury clearances or offsides passes, since injury clearances should not be counted against a player,
-            #and offside passes are not a good indicator of a player's passing ability without more information,
-            #since sometimes they are a poor decision by the passer, and sometimes they are caused by poor timing by the receiver
+            # Does not inlcude injury clearances or offsides passes, since injury clearances should not be counted against a player,
+            # and offside passes are not a good indicator of a player's passing ability without more information,
+            # since sometimes they are a poor decision by the passer, and sometimes they are caused by poor timing by the receiver
             if event['pass']['outcome'] != {'id': 74, 'name': 'Injury Clearance'} and event['pass']['outcome'] != {'id': 76, 'name': 'Pass Offside'}:
                 no_drop.append(event['id'])
                 exists = False
@@ -154,7 +141,7 @@ def xg(event):
             no_drop.append(event['id'])
             exists = False
             for key, val in shot.items():
-                #Finding the shot from the current possession and ensuring it happened after the pass
+                # Finding the shot from the current possession and ensuring it happened after the pass
                 if (key[0] == event['possession']) and (key[2] > event['index']):
                     exists = True
                     return val
@@ -163,6 +150,7 @@ def xg(event):
     else:
         return None
 
+# Build the id for the event type.
 def build(event):
     a = str(event['square'])
     b = str(event['direction'])
@@ -170,59 +158,47 @@ def build(event):
     d = str(event['related_pressure'])
     return a + b + c + d
 
+# Find the position group of a position.
 def position(pos):
-    #Dividing the positions into 8 general areas
     if type(pos) == dict:
         id = pos['id']
-        if id == 1:
-            #Goalkeeper
+        if id == 1: # Goalkeeper
             return 1
-        elif id == 2:
-            #Outside Back
+        elif id == 2: # Outside Back
             return 3
-        elif id <= 5:
-            #Center Back
+        elif id <= 5: # Center Back
             return 2
-        elif id == 6:
-            #Outside Back
+        elif id == 6: # Outside Back
             return 3
-        elif id <= 8:
-            #Winger
+        elif id <= 8: # Winger
             return 6
-        elif id <= 11:
-            #Defensive Midfielder
+        elif id <= 11: # Defensive Midfielder
             return 4
-        elif id == 12:
-            #Winger
+        elif id == 12: # Winger
             return 6
-        elif id <= 15:
-            #Center Midfielder
+        elif id <= 15: # Center Midfielder
             return 5
-        elif id <= 17:
-            #Winger
+        elif id <= 17: # Winger
             return 6
-        elif id <= 20:
-            #Attacking Midfielder
+        elif id <= 20: # Attacking Midfielder
             return 7
-        elif id == 21:
-            #Winger
+        elif id == 21: # Winger
             return 6
-        else:
-            #Forward
+        else: # Forward
             return 8
-    else:
+    else: # Not found
         return 0
 
+# Cleaning the player column
 def player_change(player):
-    #Cleaning the player column
     if type(player) == dict:
         return player['name']
     else:
         return None
 
+# Cleaning the body part section of passes and storing separately
+# Helps enable me to remove the passing data to save memory after I have stored all the information I need separately
 def body(action):
-    #Cleaning the body part section of passes and storing separately
-    #Helps enable me to remove the passing data to safe space after I have stored all the information I need separately
     if type(action) == dict:
         if 'body_part' in action:
             return action['body_part']['name']
@@ -231,11 +207,16 @@ def body(action):
     else:
         return None
 
-os.remove(r'passing_data.json')
-for file in os.listdir(r'clean_files'):
-    os.remove(r'clean_files\\' + file)
+# Remove files from previously run iterations
 
-#cleaning each file before merging them all
+if os.path.exists(r'passing_data.json'):
+    os.remove(r'passing_data.json')
+for file in os.listdir(r'clean_files'):
+    file_path = os.path.join(r'clean_files', file)
+    if os.path.isfile(file_path):
+        os.remove(r'clean_files\\' + file)
+
+# Clean each file before merging them all
 for file in os.listdir(r'statsbomb\data\events'):
     data = pd.read_json(r'statsbomb\data\\events\\' + file)
 
@@ -278,7 +259,6 @@ for file in os.listdir(r'statsbomb\data\events'):
     df['square'] = df['location'].apply(square)
     df['direction'] = df['pass'].apply(direction)
     df['alt'] = df['pass'].apply(alt)
-    df['pressure'] = df['under_pressure'].apply(pressure)
     df.apply(receipt, axis = 1)
     df['related_pressure'] = df.apply(related_pressure, axis=1)
     df.apply(shots, axis = 1)
@@ -302,23 +282,25 @@ for file in os.listdir(r'statsbomb\data\events'):
                   'interception', 'related_events', 'under_pressure'], axis = 1)
 
     df['val'] = df.apply(build, axis = 1)
-    df = df.drop(['square', 'direction', 'alt', 'pressure', 'related_pressure'], axis = 1)
+    df = df.drop(['square', 'direction', 'alt', 'related_pressure'], axis = 1)
 
-    #Save the cleaned data
+    # Save the cleaned data
+    # Although these files aren't used elsewhere, I decided as a design choice to save them to storage instead of memory to help
+    # the performance of the program.
     df.to_json(r'clean_files\\' + file[:-5] + r"_clean.json")
 
-#Ensuring no duplicate files are merged
+
 no_dup_files = set([])
-for file in os.listdir(r'clean_files'):
-    no_dup_files.add(r'clean_files\\' + file)
+for file_path in os.listdir(r'clean_files'):
+    no_dup_files.add(r'clean_files\\' + file_path)
 files = list(no_dup_files)
 
-#Merging all the files into one
+# Merging all the files into one
 def merge(files):
-    #Used recursion to avoid as much repeat reading of rows in the data frames, and it is still the same number of merges
-    #as if files were merged in a line (1 and 2, then the result of that and 3, etc.)
+    # Used recursion to avoid as much repeat reading of rows in the data frames, and it is still the same number of merges
+    # as if files were merged in a line (1 and 2, then the result of that and 3, etc.)
 
-    #base cases
+    # base cases
     if len(files) == 2:
         file_i = pd.read_json(files[0])
         file_j = pd.read_json(files[1])
@@ -327,7 +309,7 @@ def merge(files):
         return pd.read_json(files[0])
     elif len(files) == 0:
         return None
-    #recursive case
+    # recursive case
     else:
         merged1 = merge(files[:int(len(files)/2)])
         merged2 = merge(files[int(len(files)/2):])
